add javadoc to all classes
