+++ lib/src/main/java/im/status/keycard/globalplatform/Session.java
@@ -1,19 +1,36 @@
 package im.status.keycard.globalplatform;
 
+/**
+ * SCP02 Session.
+ */
 public class Session {
-    private SCP02Keys keys;
-    private byte[] cardChallenge;
+  private SCP02Keys keys;
+  private byte[] cardChallenge;
 
-    public Session(SCP02Keys keys, byte[] cardChallenge) {
-        this.keys = keys;
-        this.cardChallenge = cardChallenge;
-    }
+  /**
+   * Constructs the SCP02 session.
+   *
+   * @param keys the session keys
+   * @param cardChallenge the card challenge
+   */
+  public Session(SCP02Keys keys, byte[] cardChallenge) {
+    this.keys = keys;
+    this.cardChallenge = cardChallenge;
+  }
 
-    public SCP02Keys getKeys() {
-        return keys;
-    }
+  /**
+   * The SCP02 keys
+   * @return SCP02 keys
+   */
+  public SCP02Keys getKeys() {
+    return keys;
+  }
 
-    public byte[] getCardChallenge() {
-        return cardChallenge;
-    }
+  /**
+   * The card challenge
+   * @return card challenge
+   */
+  public byte[] getCardChallenge() {
+    return cardChallenge;
+  }
 }
