+++ lib/src/main/java/im/status/keycard/io/CardListener.java
@@ -1,6 +1,18 @@
 package im.status.keycard.io;
 
+/**
+ * Listener for card connection events.
+ */
 public interface CardListener {
-    void onConnected(CardChannel channel);
-    void onDisconnected();
+  /**
+   * Executes when the card channel is connected.
+   *
+   * @param channel the connected card channel
+   */
+  void onConnected(CardChannel channel);
+
+  /**
+   * Executes when a previously connected card is disconnected.
+   */
+  void onDisconnected();
 }
